# backend/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Football Website API",
    "version": "1.0.0",
    "description": "Public read API (v1) for news, matches, teams, players, competitions and transfers. All list responses are paginated. Errors use a stable { error: { code, message } } envelope."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [],
  "paths": {
...
    "/media/{id}": {
      "get": {
        "summary": "Media metadata + variants (public only when referenced by published articles)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media with variants"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Private media"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update media metadata alt/caption/credit (owner or editor)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated media"
          }
        }
      },
      "delete": {
        "summary": "Safe delete (409 while referenced by articles; otherwise storage+rows removed)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "409": {
            "description": "Referenced media"
          }
        }
      }
    },
    "/media/{id}/variants": {
      "get": {
        "summary": "Variant list with width/height/format/path/URL/size (same visibility as media)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Variants"
          }
        }
      }
    },
    "/media/orphans/list": {
      "get": {
        "summary": "Orphan media without article references (editor+)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Orphan media rows"
          }
        }
      }
    },
    "/media/upload": {
      "post": {
        "summary": "Upload image (author+; validates MIME/magic/dims, generates large/medium/small/thumbnail variants)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay"
          },
          "201": {
            "description": "Canonical media + variants"
          },
          "400": {
            "description": "Invalid file"
          },
          "413": {
            "description": "Too large"
          },
          "415": {
            "description": "Unsupported type"
          },
          "422": {
            "description": "Corrupt image"
          },
          "429": {
            "description": "Upload rate limited"
          }
        }
      }
    },
    "/news": {
      "get": {
        "summary": "List published news (published status + published_at <= now only)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            },
            "description": "Clamped to server maximum (default 100)"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "news",
                "breaking_news",
                "transfer",
                "match_report",
                "analysis",
                "opinion"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "player",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "competition",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "match",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "breaking",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated published articles"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/news/{slug}": {
      "get": {
        "summary": "Get published article by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article"
          },
          "400": {
            "description": "Invalid slug"
          },
          "404": {
            "description": "Not found (includes drafts)"
          }
        }
      }
    },
    "/news/breaking": {
      "get": {
        "summary": "Breaking news feed (published only)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated breaking articles"
          }
        }
      }
    },
    "/news/latest": {
      "get": {
        "summary": "Latest news feed (published only)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated latest articles"
          }
        }
      }
    },
...
    "/players": {
      "get": {
        "summary": "List players",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "position",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nationality",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated players"
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
    "/players/{slug}": {
      "get": {
        "summary": "Get player by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/players/{slug}/details": {
      "get": {
        "summary": "Player page (history, transfers, articles, recent statistics)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player details"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}